refactor: move into directories
