+++ src/types/messages.ts
@@ -1,0 +1,4 @@
+export enum MessageTypeUrl {
+  COSMOS_MSG_SEND = '/cosmos.bank.v1beta1.MsgSend',
+  LEND_MSG_DEPOSIT = '/kava.hard.v1beta1.MsgDeposit',
+}
